Tighten onboarding review fixes
