+++ apps/web/src/lib/auth/post-login-onboarding.ts
@@ -1,0 +1,11 @@
+export function shouldRedirectToOnboardingAfterLogin(opts: {
+	returnUrl: string;
+	onboardingComplete: boolean | null;
+	createdPersonalTeam: boolean;
+}) {
+	return (
+		opts.returnUrl === "/" &&
+		opts.createdPersonalTeam &&
+		opts.onboardingComplete !== true
+	);
+}
